temporary patch as navigation options is not working
